+++ CHANGELOG.md
@@ -75,7 +75,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- 
+- budget pie chart
 
 ### Changed